# Conflicts:
#	pom.xml
#	src/test/java/com/example/auctionservicesaplication/controller/AuctionControllerTest.java
#	src/test/java/com/example/auctionservicesaplication/service/AuctionServiceTest.java
#	src/test/java/com/example/auctionservicesaplication/service/BidServiceTest.java
#	src/test/java/com/example/auctionservicesaplication/service/CategoryServiceImplTest.java
#	src/test/java/com/example/auctionservicesaplication/service/UserServiceTest.java

+++ src/main/java/com/example/auctionservicesaplication/model/Auction.java
@@ -18,7 +18,7 @@ public class Auction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "auction_id")
-    private Long id;
+    private BigDecimal id;
 
     @ManyToOne
     @JoinColumn(name = "seller_id")  // adjust the column name accordingly
